Make action.h/.cpp into a module
