+++ src/gamemode.h
@@ -7,7 +7,8 @@
 
 #include "enums.h"
 
-enum action_id : int;
+import action_id_m;
+
 struct special_game;
 
 std::string special_game_name( special_game_id id );
